# JayRide_TechChallenge_API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "JayRide_TechChallenge_API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Candidate": {
      "get": {
        "tags": [
          "Candidates"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CandidateModel"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/Location": {
      "get": {
        "tags": [
          "Candidates"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/Listings": {
      "get": {
        "tags": [
          "Candidates"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/QuoteRequestModel"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "CandidateModel": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    },
    "Listing": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "pricePerPassenger": {
          "format": "float",
          "type": "number"
        },
        "vehicleType": {
          "$ref": "#/definitions/Vehicletype"
        },
        "totalPrice": {
          "format": "float",
          "type": "number"
        }
      }
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
    "QuoteRequestModel": {
      "type": "object",
      "properties": {
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "listings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Listing"
          }
        }
      }
    },
    "Vehicletype": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "maxPassengers": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}